API Reference

API documentation

DNDSparse

JuliaDB.DNDSparse — Type.

A distributed NDSparse datastructure. Can be constructed using loadfiles, ingest or distribute

source

Loading data

JuliaDB.ingest — Function.

ingest(files::Union{AbstractVector,String}, outputdir::AbstractString; <options>...)

ingests data from CSV files into JuliaDB. Stores the metadata and index in a directory outputdir. Creates outputdir if it doesn't exist.

All keyword arguments are passed to loadfiles. delim is a keyword argument to ingest – this is the delimiter in CSV reading.

Equivalent to calling loadfiles and then save on the result of loadfiles. See also loadfiles and save

source

JuliaDB.ingest! — Function.

ingest!(files::Union{AbstractVector,String}, inputdir::AbstractString, outpudir=inputdir; delim=',', chunks=1, <options>...)

load data from files and add it to data stored in inputdir. If outputdir is specified, the resulting table will be written to this directory. By default, this will create a single chunk and append it. Pass chunks argument to specify how many chunks the new data should be loaded as.

See also ingest

source

JuliaDB.loadfiles — Function.

loadfiles(files::Union{AbstractVector,String}, delim = ','; <options>)

Load a collection of CSV files into a DNDSparse, where files is either a vector of file paths, or the path of a directory containing files to load.

Arguments:

usecache::Bool: use cached metadata from previous loads while loading the files. Set this to false if you are changing other options.

All other arguments options are the same as those listed in ingest.

See also ingest.

source

Saving and Loading tables

Saving an existing DNDSparse can be accomplished through the use of the save function. The save function has the following help string:

Dagger.save — Function.

save(t::DNDSparse, outputdir::AbstractString)

Saves a DNDSparse to disk. This function blocks till all chunks have been computed and saved. Saved data can be loaded with load.

See also ingest, load

source

Loading a previously saved DNDSparse from disk can be accomplished through use of the load function. The load function has the following help string:

Dagger.load — Function.

load(dir::AbstractString; tomemory)

Load a saved DNDSparse from dir directory. Data can be saved using ingest or save functions. If tomemory option is true, then data is loaded into memory rather than mmapped.

See also ingest, save

source

distributing an NDSparse

Dagger.distribute — Function.

distribute(t::Table, chunks)

Distribute a table in chunks pieces. Equivalent to table(t, chunks=chunks).

source

distribute(itable::NDSparse, rowgroups::AbstractArray)

Distributes an NDSparse object into a DNDSparse by splitting it up into chunks of rowgroups elements. rowgroups is a vector specifying the number of rows in the chunks.

Returns a DNDSparse.

source

distribute(itable::NDSparse, nchunks::Int=nworkers())

Distributes an NDSparse object into a DNDSparse of nchunks chunks of approximately equal size.

Returns a DNDSparse.

source

Compute and collect

Operations in JuliaDB are out-of-core in nature. They return DNDSparse objects which can contain parts that are not yet evaluated. compute and collect are ways to force evaluation.

Dagger.compute — Method.

compute(t::DNDSparse; allowoverlap, closed)

Computes any delayed-evaluations in the DNDSparse. The computed data is left on the worker processes. Subsequent operations on the results will reuse the chunks.

If allowoverlap is false then the computed data is re-sorted if required to have no chunks with overlapping index ranges if necessary.

If closed is true then the computed data is re-sorted if required to have no chunks with overlapping OR continuous boundaries.

See also collect.

Warning

compute(t) requires at least as much memory as the size of the result of the computing t. You usually don't need to do this for the whole dataset. If the result is expected to be big, try compute(save(t, "output_dir")) instead. See save for more.

source

Base.collect — Method.

collect(t::DNDSparse)

Gets distributed data in a DNDSparse t and merges it into NDSparse object

Warning

collect(t) requires at least as much memory as the size of the result of the computing t. If the result is expected to be big, try compute(save(t, "output_dir")) instead. See save for more. This data can be loaded later using load.

source

Indexing

Base.getindex — Method.

t[idx...]

Returns a DNDSparse containing only the elements of t where the given indices (idx) match. If idx has the same type as the index tuple of the t, then this is considered a scalar indexing (indexing of a single value). In this case the value itself is looked up and returned.

source

Queries

Base.Sort.select — Method.

select(t::DNDSparse, conditions::Pair...)

Filter based on index columns. Conditions are accepted as column-function pairs.

Example: select(t, 1 => x->x>10, 3 => x->x!=10 ...)

source

Base.Sort.select — Method.

select(t::DNDSparse, which...; agg)

Returns a new DNDSparse where only a subset of the index columns (specified by which) are kept.

The agg keyword argument is a function which specifies how entries with equal indices should be aggregated. If agg is unspecified, then the repeating indices are kept in the output, you can then aggregate using aggregate

source

IndexedTables.aggregate — Method.

aggregate(f, t::DNDSparse)

Combines adjacent rows with equal indices using the given 2-argument reduction function f.

source

IndexedTables.aggregate_vec — Method.

aggregate_vec(f::Function, x::DNDSparse)

Combine adjacent rows with equal indices using a function from vector to scalar, e.g. mean.

source

Base.filter — Method.

filter(f, t::DNDSparse)

Filters t removing rows for which f is false. f is passed only the data and not the index.

source

IndexedTables.convertdim — Method.

convertdim(x::DNDSparse, d::DimName, xlate; agg::Function, name)

Apply function or dictionary xlate to each index in the specified dimension. If the mapping is many-to-one, agg is used to aggregate the results. name optionally specifies a name for the new dimension. xlate must be a monotonically increasing function.

See also reducedim and aggregate

source

Base.reducedim — Method.

reducedim(f, t::DNDSparse, dims)

Remove dims dimensions from t, aggregate any rows with equal indices using 2-argument function f.

See also reducedim_vec, select and aggregate.

source

IndexedTables.reducedim_vec — Method.

reducedim_vec(f::Function, t::DNDSparse, dims)

Like reducedim, except uses a function mapping a vector of values to a scalar instead of a 2-argument scalar function.

See also reducedim and aggregate_vec.

source

Joins

IndexedTables.naturaljoin — Method.

naturaljoin(left::DNDSparse, right::DNDSparse, [op])

Returns a new DNDSparse containing only rows where the indices are present both in left AND right tables. The data columns are concatenated.

source

leftjoin{K,V}(left::DNDSparse{K,V}, right::DNDSparse)

IndexedTables.asofjoin — Method.

asofjoin(left::DNDSparse, right::DNDSparse)

Keeps the indices of left but uses the value from right corresponding to highest index less than or equal to that of left.

source

Base.merge — Method.

merge(left::DNDSparse, right::DNDSparse; agg)

Merges left and right combining rows with matching indices using agg. By default agg picks the value from right.

source