API Reference

API documentation

DTable

JuliaDB.DTable — Type.

A distributed table. Can be constructed using loadfiles, ingest or distribute

source

Loading data

JuliaDB.ingest — Function.

ingest(files::Union{AbstractVector,String}, outputdir::AbstractString; <options>...)

ingests data from CSV files into JuliaDB. Stores the metadata and index in a directory outputdir. Creates outputdir if it doesn't exist.

Arguments:

delim::Char: the delimiter to use to read the text file with data. defaults to ,

indexcols::AbstractArray: columns that are meant to act as the index for the table. Defaults to all but the last column. If datacols is set, defaults to all columns other than the data columns. If indexcols is an empty vector, an implicit index of itegers 1:n is added to the data.

datacols::AbstractArray: columns that are meant to act as the data for the table. Defaults to the last column. If indexcols is set, defaults to all columns other than the index columns.

agg::Function: aggregation function to use to combine data points with the same index. Defaults to nothing which leaves the data unaggregated (see aggregate to aggregate post-loading)). table.)

presorted::Bool: specifies if each CSV file is internally already sorted according to the specified index column. This will avoid a re-sorting.

tomemory::Bool: Load data to memory after ingesting instead of mmapping. Defaults to false.

The rest of the keyword arguments will be passed on to TextParse.csvread which is used by this function to load data from individual files.

See also loadfiles and save

source

JuliaDB.ingest! — Function.

ingest!(files::Union{AbstractVector,String}, outputdir::AbstractString; <options>...)

ingest data from files and append into data stored in outputdir. Creates outputdir if it doesn't exist. Arguments are the same as those to ingest. The index range of data in the new files should not overlap with files previously ingested.

See also ingest

source

JuliaDB.loadfiles — Function.

loadfiles(files::Union{AbstractVector,String}, delim = ','; <options>)

Load a collection of CSV files into a DTable, where files is either a vector of file paths, or the path of a directory containing files to load.

Arguments:

usecache::Bool: use cached metadata from previous loads while loading the files. Set this to false if you are changing other options.

All other arguments options are the same as those listed in ingest.

See also ingest.

source

Saving and Loading tables

Saving an existing DTable can be accomplished through the use of the save function. The save function has the following help string:

JuliaDB.save — Function.

save(t::DTable, outputdir::AbstractString)

Saves a DTable to disk. This function blocks till all files data has been computed and saved. Saved data can be loaded with load.

See also ingest, load

source

Loading a previously saved DTable from disk can be accomplished through use of the load function. The load function has the following help string:

JuliaDB.load — Function.

load(dir::AbstractString; tomemory)

Load a saved DTable from dir directory. Data can be saved using ingest or save functions. If tomemory option is true, then data is loaded into memory rather than mmapped.

See also ingest, save

source

distributing an IndexedTable

JuliaDB.distribute — Function.

distribute(itable::IndexedTable, rowgroups::AbstractArray)

Distributes an IndexedTable object into a DTable by splitting it up into chunks of rowgroups elements. rowgroups is a vector specifying the number of rows in the chunks.

Returns a DTable.

source

distribute(itable::IndexedTable, nchunks::Int=nworkers())

Distributes an IndexedTable object into a DTable of nchunks chunks of approximately equal size.

Returns a DTable.

source

Compute and gather

Operations in JuliaDB are out-of-core in nature. They return DTable objects which can contain parts that are not yet evaluated. compute and gather are ways to force evaluation.

Dagger.compute — Method.

compute(t::DTable; allowoverlap, closed)

Computes any delayed-evaluations in the DTable. The computed data is left on the worker processes. Subsequent operations on the results will reuse the chunks.

If allowoverlap is false then the computed data is re-sorted if required to have no chunks with overlapping index ranges if necessary.

If closed is true then the computed data is re-sorted if required to have no chunks with overlapping OR continuous boundaries.

See also gather.

Warning

compute(t) requires at least as much memory as the size of the result of the computing t. You usually don't need to do this for the whole dataset. If the result is expected to be big, try compute(save(t, "output_dir")) instead. See save for more.

source

Dagger.gather — Method.

gather(t::DTable)

Gets distributed data in a DTable t and merges it into IndexedTable object

Warning

gather(t) requires at least as much memory as the size of the result of the computing t. If the result is expected to be big, try compute(save(t, "output_dir")) instead. See save for more. This data can be loaded later using load.

source

Indexing

Base.getindex — Method.

t[idx...]

Returns a DTable containing only the elements of t where the given indices (idx) match. If idx has the same type as the index tuple of the t, then this is considered a scalar indexing (indexing of a single value). In this case the value itself is looked up and returned.

source

Queries

Base.Sort.select — Method.

select(t::DTable, conditions::Pair...)

Filter based on index columns. Conditions are accepted as column-function pairs.

Example: select(t, 1 => x->x>10, 3 => x->x!=10 ...)

source

Base.Sort.select — Method.

select(t::DTable, which...; agg)

Returns a new DTable where only a subset of the index columns (specified by which) are kept.

The agg keyword argument is a function which specifies how entries with equal indices should be aggregated. If agg is unspecified, then the repeating indices are kept in the output, you can then aggregate using aggregate

source

IndexedTables.aggregate — Method.

aggregate(f, t::DTable)

Combines adjacent rows with equal indices using the given 2-argument reduction function f.

source

IndexedTables.aggregate_vec — Method.

aggregate_vec(f::Function, x::DTable)

Combine adjacent rows with equal indices using a function from vector to scalar, e.g. mean.

source

Base.filter — Method.

filter(f, t::DTable)

Filters t removing rows for which f is false. f is passed only the data and not the index.

source

IndexedTables.convertdim — Method.

convertdim(x::DTable, d::DimName, xlate; agg::Function, name)

Apply function or dictionary xlate to each index in the specified dimension. If the mapping is many-to-one, agg is used to aggregate the results. name optionally specifies a name for the new dimension. xlate must be a monotonically increasing function.

See also reducedim and aggregate

source

Base.reducedim — Method.

reducedim(f, t::DTable, dims)

Remove dims dimensions from t, aggregate any rows with equal indices using 2-argument function f.

See also reducedim_vec, select and aggregate.

source

IndexedTables.reducedim_vec — Method.

reducedim_vec(f::Function, t::DTable, dims)

Like reducedim, except uses a function mapping a vector of values to a scalar instead of a 2-argument scalar function.

See also reducedim and aggregate_vec.

source

Joins

IndexedTables.naturaljoin — Method.

naturaljoin(left::DTable, right::DTable, [op])

Returns a new DTable containing only rows where the indices are present both in left AND right tables. The data columns are concatenated.

source

IndexedTables.leftjoin — Method.

leftjoin(left::DTable, right::DTable, [op::Function])

Keeps only rows with indices in left. If rows of the same index are present in right, then they are combined using op. op by default picks the value from right.

source

IndexedTables.asofjoin — Method.

asofjoin(left::DTable, right::DTable)

Keeps the indices of left but uses the value from right corresponding to highest index less than or equal to that of left.

source

Base.merge — Method.

merge(left::DTable, right::DTable; agg)

Merges left and right combining rows with matching indices using agg. By default agg picks the value from right.

source